Changed maximumEntries to correct int rather than long

+++ opendaylight/sal/api/src/main/java/org/opendaylight/controller/sal/reader/NodeTableStatistics.java
@@ -169,7 +169,7 @@ public class NodeTableStatistics implements Serializable {
     /**
      * @return the maximumEntries
      */
-    public long getMaximumEntries() {
+    public int getMaximumEntries() {
         return maximumEntries;
     }
